+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -9,6 +9,7 @@ This version of the client library defaults to the service API version `2023-07-
 ### Breaking Changes
 - Going forward this library defaults to the service API version `2023-07-31`.
 - Removed `query_fields` keyword argument from `begin_analyze_document()` and `begin_analyze_document_from_url()`.
+- Removed `kind` property from `DocumentPage`.
 
 ### Bugs Fixed